requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-crispy-forms==1.14.0
django-phonenumber-field==6.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
phonenumbers==8.12.49
Pillow==9.1.1
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
python-dateutil==1.5
python-dotenv==0.20.0
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
